requirements.txt
Flask>=2.2.2,<3
google-cloud-logging>=3.2.5,<4